refactor(institution): Move Institutions to Movements namespace
